+++ app/controllers/rails_health_checks/application_controller.rb
@@ -2,5 +2,7 @@
 
 module RailsHealthChecks
   class ApplicationController < ActionController::API
+    include Authentication
+    before_action :authenticate!
   end
 end